# rpl-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BackEnd RPL - By Dewa Permana",
        "description": "Swagger Documentation BackEnd Web RPL - Created by Dewa Permana",
        "contact": {
            "email": "dewaprmanaptr@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kategori-berita": {
            "get": {
                "tags": [
                    "KategoriBerita"
                ],
                "summary": "Ambil semua kategori berita",
                "operationId": "13f56d9c9bbd6cc34b6195bb3bc61793",
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sukses mengambil data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "kategori": {
                                                        "type": "string",
                                                        "example": "Teknologi"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "KategoriBerita"
                ],
                "summary": "Tambah Kategori Berita",
                "operationId": "7242f9bce9694ffb1b5e905d2e3913a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori"
                                ],
                                "properties": {
                                    "kategori": {
                                        "type": "string",
                                        "example": "Teknologi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori Berita Berhasil Ditambahkan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/kategori-berita/{id}": {
            "put": {
                "tags": [
                    "KategoriBerita"
                ],
                "summary": "Update kategori berita",
                "operationId": "00b5c58261b83c5864679e4199bb0404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategori berita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori"
                                ],
                                "properties": {
                                    "kategori": {
                                        "type": "string",
                                        "example": "Politik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori Berita berhasil diperbarui"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "kategori": {
                                                    "type": "string",
                                                    "example": "Politik"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "KategoriBerita"
                ],
                "summary": "Hapus kategori berita",
                "operationId": "c564bbf83591ba07f0b4f70ea8ac2342",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategori berita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori Berita berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari login (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}